Dash/requirements.txt
astroid==2.12.10
autopep8==1.7.0
Babel==2.10.3
Brotli==1.0.9
click==8.1.3
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.5.1
Flask==2.2.2
Flask-Compress==1.13
greenlet==1.1.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.3
pandas==1.5.0
pandas-stubs==1.4.4.220919
platformdirs==2.5.2
plotly==5.10.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pylint==2.15.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-i18n==0.3.9
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.41
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
types-pytz==2022.2.1.0
Werkzeug==2.2.2
wrapt==1.14.1
